+++ library/Modules/Backup/Model/Pgsql.php
@@ -1,0 +1,104 @@
+<?php
+/**
+ * This source file is part of GotCms.
+ *
+ * GotCms is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * GotCms is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License along
+ * with GotCms. If not, see <http://www.gnu.org/licenses/lgpl-3.0.html>.
+ *
+ * PHP Version >=5.3
+ *
+ * @category   Gc_Library
+ * @package    Modules
+ * @subpackage Blog\Model
+ * @author     Pierre Rambaud (GoT) <pierre.rambaud86@gmail.com>
+ * @license    GNU/LGPL http://www.gnu.org/licenses/lgpl-3.0.html
+ * @link       http://www.got-cms.com
+ */
+
+namespace Modules\Backup\Model;
+
+use Gc\Db\AbstractTable,
+    Gc\Document\Model as DocumentModel,
+    Zend\Db\Sql\Select,
+    Zend\Db\Sql\Predicate\Expression;
+
+/**
+ * Blog comment table
+ *
+ * @category   Gc_Library
+ * @package    Modules
+ * @subpackage Blog\Model
+ */
+class Pgsql extends AbstractTable
+{
+    /**
+     * Table name
+     *
+     * @var string
+     */
+    protected $_name = 'core_config_data';
+
+    /**
+     * Export function
+     *
+     * @param string $what
+     * @return string
+     */
+    public function export($what = 'structureanddata')
+    {
+        $exe = escapeshellcmd('/usr/bin/pg_dump');
+        $parameters = $this->getAdapter()->getDriver()->getConnection()->getConnectionParameters();
+
+        // Set environmental variables that pg_dump uses
+        putenv('PGPASSWORD=' . $parameters['password']);
+        putenv('PGUSER=' . $parameters['username']);
+        putenv('PGDATABASE=' . $parameters['database']);
+        if(!empty($parameters['hostname']))
+        {
+            putenv('PGHOST=' . $parameters['hostname']);
+        }
+
+        if(!empty($parameters['port']))
+        {
+            putenv('PGPORT=' . $port);
+        }
+
+        //Prepare command
+        $cmd = $exe . ' -i';
+        $cmd .= ' -Z 9';
+
+        switch ($what)
+        {
+            case 'dataonly':
+                $cmd .= ' -a';
+                $cmd .= ' --inserts';
+            break;
+
+            case 'structureonly':
+                $cmd .= ' -s';
+                $cmd .= ' -c';
+            break;
+
+            case 'all':
+                $cmd .= ' --inserts';
+                $cmd .= ' -c';
+            break;
+        }
+
+
+        // Execute command and return the output
+        exec($cmd . ' 2>&1', $output);
+
+        return implode(PHP_EOL, $output);
+    }
+}